Изменён тип поля с BigInteger на Long

+++ src/main/java/com/art/repository/FacilityRepository.java
@@ -6,21 +6,20 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigInteger;
 import java.util.List;
 
 @Repository
-public interface FacilityRepository extends JpaRepository<Facility, BigInteger> {
+public interface FacilityRepository extends JpaRepository<Facility, Long> {
 
-    Facility findById(BigInteger id);
+    Facility findById(Long id);
 
     Facility findByName(String facility);
 
     List<Facility> findAll();
 
-    List<Facility> findByIdNot(BigInteger id);
+    List<Facility> findByIdNot(Long id);
 
-    List<Facility> findByIdIn(List<BigInteger> idList);
+    List<Facility> findByIdIn(List<Long> idList);
 
     @Query(name = "InvestorsFacilities", nativeQuery = true)
     List<UserFacilities> getInvestorsFacility(Long rentorInvestorId);
